draft of decomposition of model_interfaces.py
