+++ trieste/models/models_gpflow/interface.py
@@ -1,0 +1,82 @@
+# Copyright 2020 The Trieste Contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+
+import gpflow
+import tensorflow as tf
+from gpflow.models import GPModel
+
+from ...data import Dataset
+from ...type import TensorType
+from ..optimizer import Optimizer
+from ..interfaces import ProbabilisticModel
+from .utils import module_deepcopy
+
+
+class GPflowPredictor(ProbabilisticModel, tf.Module, ABC):
+    """A trainable wrapper for a GPflow Gaussian process model."""
+
+    def __init__(self, optimizer: Optimizer | None = None):
+        """
+        :param optimizer: The optimizer with which to train the model. Defaults to
+            :class:`~trieste.models.optimizer.Optimizer` with :class:`~gpflow.optimizers.Scipy`.
+        """
+        super().__init__()
+
+        if optimizer is None:
+            optimizer = Optimizer(gpflow.optimizers.Scipy())
+
+        self._optimizer = optimizer
+
+    @property
+    def optimizer(self) -> Optimizer:
+        """The optimizer with which to train the model."""
+        return self._optimizer
+
+    @property
+    @abstractmethod
+    def model(self) -> GPModel:
+        """The underlying GPflow model."""
+
+    def predict(self, query_points: TensorType) -> tuple[TensorType, TensorType]:
+        return self.model.predict_f(query_points)
+
+    def predict_joint(self, query_points: TensorType) -> tuple[TensorType, TensorType]:
+        return self.model.predict_f(query_points, full_cov=True)
+
+    def sample(self, query_points: TensorType, num_samples: int) -> TensorType:
+        return self.model.predict_f_samples(query_points, num_samples)
+
+    def predict_y(self, query_points: TensorType) -> tuple[TensorType, TensorType]:
+        return self.model.predict_y(query_points)
+
+    def get_kernel(self) -> gpflow.kernels.Kernel:
+        """
+        Return the kernel of the model.
+
+        :return: The kernel.
+        """
+        return self.model.kernel
+
+    def optimize(self, dataset: Dataset) -> None:
+        """
+        Optimize the model with the specified `dataset`.
+
+        :param dataset: The data with which to optimize the `model`.
+        """
+        self.optimizer.optimize(self.model, dataset)
+
+    __deepcopy__ = module_deepcopy
